Move desugar unit tests to desugar
In future this class will become more complex, so
at that point these will also be split up.
