+++ src/desugar/node.rs
@@ -354,3 +354,458 @@ pub fn desugar_node(ast: &AST, imp: &mut Imports, state: &State) -> DesugarResul
 
     Ok(core)
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::common::position::Position;
+    use crate::core::construct::Core;
+    use crate::desugar::desugar;
+    use crate::parse::ast::AST;
+    use crate::parse::ast::Node;
+
+    macro_rules! to_pos_unboxed {
+        ($node:expr) => {{
+            AST { pos: Position::default(), node: $node }
+        }};
+    }
+
+    macro_rules! to_pos {
+        ($node:expr) => {{
+            Box::from(to_pos_unboxed!($node))
+        }};
+    }
+
+    #[test]
+    fn break_verify() {
+        let _break = to_pos!(Node::Break);
+        assert_eq!(desugar(&_break).unwrap(), Core::Break);
+    }
+
+    #[test]
+    fn continue_verify() {
+        let _continue = to_pos!(Node::Continue);
+        assert_eq!(desugar(&_continue).unwrap(), Core::Continue);
+    }
+
+    #[test]
+    fn pass_verify() {
+        let pass = to_pos!(Node::Pass);
+        assert_eq!(desugar(&pass).unwrap(), Core::Pass);
+    }
+
+    #[test]
+    fn print_verify() {
+        let expr = to_pos!(Node::Str { lit: String::from("a"), expressions: vec![] });
+        let print_stmt = to_pos!(Node::Print { expr });
+        assert_eq!(desugar(&print_stmt).unwrap(), Core::Print {
+            expr: Box::from(Core::Str { string: String::from("a") })
+        });
+    }
+
+    #[test]
+    fn return_verify() {
+        let expr = to_pos!(Node::Str { lit: String::from("a"), expressions: vec![] });
+        let print_stmt = to_pos!(Node::Return { expr });
+
+        assert_eq!(desugar(&print_stmt).unwrap(), Core::Return {
+            expr: Box::from(Core::Str { string: String::from("a") })
+        });
+    }
+
+    #[test]
+    fn return_empty_verify() {
+        let print_stmt = to_pos!(Node::ReturnEmpty);
+        assert_eq!(desugar(&print_stmt).unwrap(), Core::Return { expr: Box::from(Core::None) });
+    }
+
+    #[test]
+    fn init_verify() {
+        let _break = to_pos!(Node::Init);
+        assert_eq!(desugar(&_break).unwrap(), Core::Id { lit: String::from("init") });
+    }
+
+    #[test]
+    fn self_verify() {
+        let _break = to_pos!(Node::_Self);
+        assert_eq!(desugar(&_break).unwrap(), Core::Id { lit: String::from("self") });
+    }
+
+    #[test]
+    fn import_verify() {
+        let _break = to_pos!(Node::Import {
+        import: vec![to_pos_unboxed!(Node::Id { lit: String::from("a") })],
+        aliases:    vec![to_pos_unboxed!(Node::Id { lit: String::from("b") })]
+    });
+
+        assert_eq!(desugar(&_break).unwrap(), Core::ImportAs {
+            imports: vec![Core::Id { lit: String::from("a") }],
+            alias: vec![Core::Id { lit: String::from("b") }],
+        });
+    }
+
+    #[test]
+    fn from_import_as_verify() {
+        let _break = to_pos!(Node::FromImport {
+        id:     to_pos!(Node::Id { lit: String::from("f") }),
+        import: to_pos!(Node::Import {
+            import: vec![to_pos_unboxed!(Node::Id { lit: String::from("a") })],
+            aliases:    vec![to_pos_unboxed!(Node::Id { lit: String::from("b") })]
+        })
+    });
+
+        assert_eq!(desugar(&_break).unwrap(), Core::FromImport {
+            from: Box::from(Core::Id { lit: String::from("f") }),
+            import: Box::from(Core::ImportAs {
+                imports: vec![Core::Id { lit: String::from("a") }],
+                alias: vec![Core::Id { lit: String::from("b") }],
+            }),
+        });
+    }
+
+    #[test]
+    fn raises_empty_verify() {
+        let type_def = to_pos!(Node::Raises {
+        expr_or_stmt: Box::from(to_pos!(Node::Id { lit: String::from("a") })),
+        errors:       vec![]
+    });
+        assert_eq!(desugar(&type_def).unwrap(), Core::Id { lit: String::from("a") });
+    }
+    macro_rules! verify_op {
+    ($op:ident) => {{
+        let add_op = to_pos!(Node::$op);
+        let core = desugar(&add_op).unwrap();
+        assert_eq!(core, Core::$op);
+    }};
+}
+
+    macro_rules! verify {
+    ($ast:ident) => {{
+        let left = Node::Id { lit: String::from("left") };
+        let right = Node::Id { lit: String::from("right") };
+        let add_node = to_pos!(Node::$ast { left: to_pos!(left), right: to_pos!(right) });
+
+        let (left, right) = match desugar(&add_node) {
+            Ok(Core::$ast { left, right }) => (left, right),
+            other => panic!("Expected binary operation but was {:?}", other)
+        };
+
+        assert_eq!(*left, Core::Id { lit: String::from("left") });
+        assert_eq!(*right, Core::Id { lit: String::from("right") });
+    }};
+}
+
+    macro_rules! verify_unary {
+    ($ast:ident) => {{
+        let expr = to_pos!(Node::Id { lit: String::from("expression") });
+        let add_node = to_pos!(Node::$ast { expr });
+
+        let expr_des = match desugar(&add_node) {
+            Ok(Core::$ast { expr }) => expr,
+            other => panic!("Expected unary operation but was {:?}", other)
+        };
+
+        assert_eq!(*expr_des, Core::Id { lit: String::from("expression") });
+    }};
+}
+
+    #[test]
+    fn add_verify() {
+        verify!(Add);
+    }
+
+    #[test]
+    fn sub_verify() {
+        verify!(Sub);
+    }
+
+    #[test]
+    fn mul_verify() {
+        verify!(Mul);
+    }
+
+    #[test]
+    fn div_verify() {
+        verify!(Div);
+    }
+
+    #[test]
+    fn mod_verify() {
+        verify!(Mod);
+    }
+
+    #[test]
+    fn pow_verify() {
+        verify!(Pow);
+    }
+
+    #[test]
+    fn add_unary_verify() {
+        verify_unary!(AddU);
+    }
+
+    #[test]
+    fn sub_unary_verify() {
+        verify_unary!(SubU);
+    }
+
+    #[test]
+    fn sqrt_verify() {
+        verify_unary!(Sqrt);
+    }
+
+    #[test]
+    fn le_verify() {
+        verify!(Le);
+    }
+
+    #[test]
+    fn leq_verify() {
+        verify!(Leq);
+    }
+
+    #[test]
+    fn ge_verify() {
+        verify!(Ge);
+    }
+
+    #[test]
+    fn geq_verify() {
+        verify!(Geq);
+    }
+
+    #[test]
+    fn neq_verify() {
+        verify!(Neq);
+    }
+
+    #[test]
+    fn is_verify() {
+        verify!(Is);
+    }
+
+    #[test]
+    fn not_verify() {
+        verify_unary!(Not);
+    }
+
+    #[test]
+    fn and_verify() {
+        verify!(And);
+    }
+
+    #[test]
+    fn or_verify() {
+        verify!(Or);
+    }
+
+    #[test]
+    fn add_op_verify() {
+        verify_op!(AddOp);
+    }
+
+    #[test]
+    fn sub_op_verify() {
+        verify_op!(SubOp);
+    }
+
+    #[test]
+    fn sqrt_op_verify() {
+        let sqrt_node = to_pos!(Node::SqrtOp);
+        let result = desugar(&sqrt_node);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn mul_op_verify() {
+        verify_op!(MulOp);
+    }
+
+    #[test]
+    fn div_op_verify() {
+        verify_op!(DivOp);
+    }
+
+    #[test]
+    fn pow_op_verify() {
+        verify_op!(PowOp);
+    }
+
+    #[test]
+    fn mod_op_verify() {
+        verify_op!(ModOp);
+    }
+
+    #[test]
+    fn eq_op_verify() {
+        verify_op!(EqOp);
+    }
+
+    #[test]
+    fn le_op_verify() {
+        verify_op!(LeOp);
+    }
+
+    #[test]
+    fn ge_op_verify() {
+        verify_op!(GeOp);
+    }
+
+    #[test]
+    fn tuple_verify() {
+        let elements = vec![
+            to_pos_unboxed!(Node::ENum { num: String::from("a"), exp: String::from("100") }),
+            to_pos_unboxed!(Node::Real { lit: String::from("3000.5") }),
+        ];
+        let tuple = to_pos!(Node::Tuple { elements });
+        let core = desugar(&tuple);
+
+        let core_elements = match core {
+            Ok(Core::Tuple { elements }) => elements,
+            other => panic!("Expected tuple but got {:?}", other)
+        };
+
+        assert_eq!(core_elements[0], Core::ENum { num: String::from("a"), exp: String::from("100") });
+        assert_eq!(core_elements[1], Core::Float { float: String::from("3000.5") });
+    }
+
+    #[test]
+    fn set_verify() {
+        let elements = vec![
+            to_pos_unboxed!(Node::Id { lit: String::from("a") }),
+            to_pos_unboxed!(Node::Bool { lit: true }),
+        ];
+        let set = to_pos!(Node::Set { elements });
+        let core = desugar(&set);
+
+        let core_elements = match core {
+            Ok(Core::Set { elements }) => elements,
+            other => panic!("Expected set but got {:?}", other)
+        };
+
+        assert_eq!(core_elements[0], Core::Id { lit: String::from("a") });
+        assert_eq!(core_elements[1], Core::Bool { boolean: true });
+    }
+
+    #[test]
+    fn list_verify() {
+        let elements = vec![
+            to_pos_unboxed!(Node::ENum { num: String::from("a"), exp: String::from("100") }),
+            to_pos_unboxed!(Node::Real { lit: String::from("3000.5") }),
+        ];
+        let tuple = to_pos!(Node::List { elements });
+        let core = desugar(&tuple);
+
+        let core_elements = match core {
+            Ok(Core::List { elements }) => elements,
+            other => panic!("Expected tuple but got {:?}", other)
+        };
+
+        assert_eq!(core_elements[0], Core::ENum { num: String::from("a"), exp: String::from("100") });
+        assert_eq!(core_elements[1], Core::Float { float: String::from("3000.5") });
+    }
+
+    #[test]
+    fn set_builder_verify() {
+        let item = to_pos!(Node::Id { lit: String::from("a") });
+        let conditions = vec![];
+        let list_builder = to_pos!(Node::SetBuilder { item, conditions });
+
+        let desugar_result = desugar(&list_builder);
+        assert!(desugar_result.is_err());
+    }
+
+    #[test]
+    fn list_builder_verify() {
+        let item = to_pos!(Node::Id { lit: String::from("a") });
+        let conditions = vec![];
+        let list_builder = to_pos!(Node::ListBuilder { item, conditions });
+
+        let desugar_result = desugar(&list_builder);
+        assert!(desugar_result.is_err());
+    }
+
+    #[test]
+    fn with_verify() {
+        let resource = to_pos!(Node::Id { lit: String::from("my_resource") });
+        let alias = Some((to_pos!(Node::Id { lit: String::from("other") }), false, None));
+        let expr = to_pos!(Node::Int { lit: String::from("9") });
+        let with = to_pos!(Node::With { resource, alias, expr });
+
+        let (resource, alias, expr) = match desugar(&with) {
+            Ok(Core::WithAs { resource, alias, expr }) => (resource, alias, expr),
+            other => panic!("Expected with as but was {:?}", other)
+        };
+
+        assert_eq!(*resource, Core::Id { lit: String::from("my_resource") });
+        assert_eq!(*alias, Core::Id { lit: String::from("other") });
+        assert_eq!(*expr, Core::Int { int: String::from("9") });
+    }
+
+    #[test]
+    fn with_no_as_verify() {
+        let resource = to_pos!(Node::Id { lit: String::from("other") });
+        let expr = to_pos!(Node::Int { lit: String::from("2341") });
+        let with = to_pos!(Node::With { resource, alias: None, expr });
+
+        let (resource, expr) = match desugar(&with) {
+            Ok(Core::With { resource, expr }) => (resource, expr),
+            other => panic!("Expected with but was {:?}", other)
+        };
+
+        assert_eq!(*resource, Core::Id { lit: String::from("other") });
+        assert_eq!(*expr, Core::Int { int: String::from("2341") });
+    }
+
+    #[test]
+    fn handle_empty_verify() {
+        let expr_or_stmt = to_pos!(Node::Id { lit: String::from("my_fun") });
+        let handle = to_pos!(Node::Handle { expr_or_stmt, cases: vec![] });
+
+        let (setup, _try, except) = match desugar(&handle) {
+            Ok(Core::TryExcept { setup, attempt, except }) =>
+                (setup.clone(), attempt.clone(), except.clone()),
+            other => panic!("Expected try except but was {:?}", other)
+        };
+
+        assert_eq!(setup, None);
+        assert_eq!(*_try, Core::Id { lit: String::from("my_fun") });
+        assert!(except.is_empty());
+    }
+
+    #[test]
+    fn handle_verify() {
+        let expr_or_stmt = to_pos!(Node::Id { lit: String::from("my_fun") });
+        let cond = to_pos!(Node::ExpressionType {
+        expr:    to_pos!(Node::Id { lit: String::from("err") }),
+        mutable: false,
+        ty:      Some(to_pos!(Node::Type {
+            id:       to_pos!(Node::Id { lit: String::from("my_type") }),
+            generics: vec![]
+        }))
+    });
+        let body = to_pos!(Node::Int { lit: String::from("9999") });
+        let case = to_pos_unboxed!(Node::Case { cond, body });
+        let handle = to_pos!(Node::Handle { expr_or_stmt, cases: vec![case] });
+
+        let (setup, _try, except) = match desugar(&handle) {
+            Ok(Core::TryExcept { setup, attempt, except }) =>
+                (setup.clone(), attempt.clone(), except.clone()),
+            other => panic!("Expected try except but was {:?}", other)
+        };
+
+        assert_eq!(setup, None);
+        assert_eq!(*_try, Core::Id { lit: String::from("my_fun") });
+        assert_eq!(except.len(), 1);
+        match &except[0] {
+            Core::Except { id, class, body } => {
+                assert_eq!(*id, Box::from(Core::Id { lit: String::from("err") }));
+                assert_eq!(
+                    *class,
+                    Some(Box::from(Core::Type { lit: String::from("my_type"), generics: vec![] }))
+                );
+                assert_eq!(*body, Box::from(Core::Int { int: String::from("9999") }));
+            }
+            other => panic!("Expected except case but was {:?}", other)
+        }
+    }
+}
