--- tests/desugar/mod.rs
@@ -1,11 +1,0 @@
-#[macro_use]
-mod util;
-
-pub mod class;
-pub mod collection;
-pub mod control_flow;
-pub mod definition;
-pub mod error;
-pub mod expression_and_statements;
-pub mod operation;
-pub mod types;
