--- tests/desugar/types.rs
@@ -1,1 +1,0 @@
-
