+++ tests/main.rs
@@ -11,7 +11,6 @@ mod common;
 
 mod check;
 mod core;
-mod desugar;
 mod output;
 mod parse;
